Conflicts:
	common/src/main/scala/com/precog/common/json/CPath.scala
	daze/src/main/scala/com/precog/daze/Evaluator.scala
	util/src/main/scala/com/precog/util/ByteBufferPool.scala
	yggdrasil/src/main/scala/com/precog/yggdrasil/CValue.scala
	yggdrasil/src/main/scala/com/precog/yggdrasil/SValue.scala
	yggdrasil/src/main/scala/com/precog/yggdrasil/Schema.scala
	yggdrasil/src/main/scala/com/precog/yggdrasil/TableModule.scala
	yggdrasil/src/main/scala/com/precog/yggdrasil/jdbm3/CValueSerializer.scala
	yggdrasil/src/main/scala/com/precog/yggdrasil/jdbm3/Codec.scala
	yggdrasil/src/main/scala/com/precog/yggdrasil/jdbm3/ColumnCodec.scala
	yggdrasil/src/main/scala/com/precog/yggdrasil/jdbm3/JDBMProjection.scala
	yggdrasil/src/main/scala/com/precog/yggdrasil/jdbm3/JDBMRawSortProjection.scala
	yggdrasil/src/main/scala/com/precog/yggdrasil/jdbm3/JDBMSlice.scala
	yggdrasil/src/main/scala/com/precog/yggdrasil/jdbm3/RowFormat.scala
	yggdrasil/src/main/scala/com/precog/yggdrasil/jdbm3/SortingSerialization.scala
	yggdrasil/src/main/scala/com/precog/yggdrasil/jdbm3/StdCodecs.scala
	yggdrasil/src/main/scala/com/precog/yggdrasil/leveldb/LevelDBProjectionOps.scala
	yggdrasil/src/main/scala/com/precog/yggdrasil/table/ArrayColumn.scala
	yggdrasil/src/main/scala/com/precog/yggdrasil/table/BlockStoreColumnarTableModule.scala
	yggdrasil/src/main/scala/com/precog/yggdrasil/table/Column.scala
	yggdrasil/src/main/scala/com/precog/yggdrasil/table/ColumnarTableModule.scala
	yggdrasil/src/main/scala/com/precog/yggdrasil/table/Slice.scala
	yggdrasil/src/main/scala/com/precog/yggdrasil/table/cf/Util.scala
	yggdrasil/src/test/scala/com/precog/yggdrasil/jdbm3/CValueGenerators.scala
	yggdrasil/src/test/scala/com/precog/yggdrasil/jdbm3/CodecSpec.scala
	yggdrasil/src/test/scala/com/precog/yggdrasil/jdbm3/RowFormatSpec.scala
	yggdrasil/src/test/scala/com/precog/yggdrasil/table/BlockSortSpec.scala

+++ regen-jdbm-data.sh
@@ -19,28 +19,77 @@
 ## 
 #!/bin/bash
 
+SRCDIR=muspelheim/src/test/resources/test_data
+OWNERTOKEN=C18ED787-BF07-4097-B819-0415C759C8D5
+
+function usage() {
+    echo "Usage: `basename $0` [-n] [-t <owner token>] [-s <source directory>] <target data directory>" >&2
+    echo "  -n : don't wipe existing data" >&2
+    echo "  -t : Specify the owner token (defaults to $OWNERTOKEN)" >&2
+    echo "  -s : Specify the source directory (defaults to `dirname $0`/$SRCDIR)" >&2
+    exit 1
+}
+
+VERSION=`git describe`
+
+while getopts "nt:s:" OPTNAME; do
+    case $OPTNAME in
+        t)
+            OWNERTOKEN=$OPTARG
+            ;;
+        s)
+            [ -d $OPTARG ] || {
+                echo "Could not open source directory: $OPTARG" >&2
+                exit 2
+            }
+            SRCDIR=$OPTARG
+            ;;
+        n)
+            echo "Not wiping existing data"
+            DONTWIPE=true
+            ;;
+        \?)
+            usage
+            ;;
+    esac
+done
+
+shift $(( $OPTIND - 1 ))
+
 if [[ $# != 1 ]]; then
-    echo "Usage: `basename $0` <target data directory>"
-    echo "  For now target is normally pandora/dist/data-jdbm/data/"
-    exit
+    usage
 fi
 
-SRCDIR=muspelheim/src/test/resources/test_data
-DATADIR=$1
+[ -d $1 ] || {
+    echo "Could not open target directory: $1" >&2
+    exit 2
+}
+
+DATADIR=$(cd $1; pwd)
 
 cd `dirname $0`
 
+echo "Loading data from $SRCDIR with owner token $OWNERTOKEN"
+
+if [ ! -d $SRCDIR -o ! -d $DATADIR ]; then
+    echo "Source or dest dir does not exist!" >&2
+    exit 2
+fi
+
 pushd $SRCDIR > /dev/null
 SOURCES=""
-for source in `find . -name '*.json'`; do
+for source in `find -L . -name '*.json'`; do
     DPATH=`echo $source | sed -e 's/^\.\(.*\)\.json/\1\//'`
     DATAFILE=${source/\.\//}
     SOURCES="$SOURCES $DPATH=$SRCDIR/$DATAFILE"
 done
 popd > /dev/null
 
-[ -f yggdrasil/target/yggdrasil-assembly-2.0.0-SNAPSHOT.jar ] || sbt yggdrasil/assembly
+[ -f yggdrasil/target/yggdrasil-assembly-$VERSION.jar ] || {
+    echo "Error: you must build yggdrasil/assembly before running tasks that extract data"
+    exit 2
+}
 
-rm -rf $DATADIR/*
+[ -z "$DONTWIPE" ] && rm -rf $DATADIR/*
 
-java -cp yggdrasil/target/yggdrasil-assembly-2.0.0-SNAPSHOT.jar com.precog.yggdrasil.util.YggUtils import -t C18ED787-BF07-4097-B819-0415C759C8D5 -s $DATADIR $SOURCES
+java -Xmx1G -cp yggdrasil/target/yggdrasil-assembly-$VERSION.jar com.precog.yggdrasil.util.YggUtils import -t $OWNERTOKEN -s $DATADIR $SOURCES
